# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TechPath Jobs & Internships API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        headline:
          type: string
        phone:
          type: string
        location:
          type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        portfolioUrl:
          type: string
        summary:
          type: string
        skills:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            $ref: "#/components/schemas/Education"
        experience:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        certifications:
          type: array
          items:
            $ref: "#/components/schemas/Certification"
        languages:
          type: array
          items:
            type: string
        preferredRoles:
          type: array
          items:
            type: string
        preferredLocations:
          type: array
          items:
            type: string
        expectedSalary:
          type: number
        availableFrom:
          type: string
          format: date
        openToRemote:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, skills, education, experience, projects, certifications]
    ProfileInput:
      type: object
      properties:
        headline:
          type: string
        phone:
          type: string
        location:
          type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        portfolioUrl:
          type: string
        summary:
          type: string
        skills:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            $ref: "#/components/schemas/Education"
        experience:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        certifications:
          type: array
          items:
            $ref: "#/components/schemas/Certification"
        languages:
          type: array
          items:
            type: string
        preferredRoles:
          type: array
          items:
            type: string
        preferredLocations:
          type: array
          items:
            type: string
        expectedSalary:
          type: number
        availableFrom:
          type: string
          format: date
        openToRemote:
          type: boolean
      required: [skills, education, experience, projects, certifications]
    Education:
      type: object
      properties:
        institution:
          type: string
        degree:
          type: string
        field:
          type: string
        startYear:
          type: integer
        endYear:
          type: integer
        gpa:
          type: string
        achievements:
          type: string
      required: [institution, degree, field, startYear]
    Experience:
      type: object
      properties:
        company:
          type: string
        role:
          type: string
        location:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        current:
          type: boolean
        description:
          type: string
        techStack:
          type: array
          items:
            type: string
      required: [company, role, startDate]
    Project:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        techStack:
          type: array
          items:
            type: string
        url:
          type: string
        githubUrl:
          type: string
        startDate:
          type: string
        endDate:
          type: string
      required: [name, description]
    Certification:
      type: object
      properties:
        name:
          type: string
        issuer:
          type: string
        issueDate:
          type: string
        expiryDate:
          type: string
        credentialUrl:
          type: string
      required: [name, issuer]
...