# lib/api-zod/src/generated/types/education.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechPath Jobs & Internships API
 * OpenAPI spec version: 0.1.0
 */

export interface Education {
  institution: string;
  degree: string;
  field: string;
  startYear: number;
  endYear?: number;
  gpa?: string;
  achievements?: string;
}