# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechPath Jobs & Internships API
 * OpenAPI spec version: 0.1.0
 */
import type { Certification } from "./certification";
import type { Education } from "./education";
import type { Experience } from "./experience";
import type { Project } from "./project";

export interface Profile {
  id: number;
  userId: number;
  headline?: string;
  phone?: string;
  location?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  portfolioUrl?: string;
  summary?: string;
  skills: string[];
  education: Education[];
  experience: Experience[];
  projects: Project[];
  certifications: Certification[];
  languages?: string[];
  preferredRoles?: string[];
  preferredLocations?: string[];
  expectedSalary?: number;
  availableFrom?: Date;
  openToRemote?: boolean;
  updatedAt?: Date;
}